pyproject.toml
[build-system]
requires=["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "teyvat-lore-graph"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "neo4j>=6.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]

[tool.setuptools.packages.find]
where = ["src"]
